fix false positive len_zero in is_empty()
